pyproject.toml
[project]
name = "openshell-deep-agent"
version = "0.1.0"
description = "General-purpose deep agent using OpenShell as the on-prem sandbox provider"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.4.11",
    "langchain>=1.2.12",
    "langchain-nvidia-ai-endpoints>=1.2.1",
    "langgraph-cli[inmem]>=0.4.18",
    "openshell>=0.0.7",
    "PyYAML>=6.0",
    "langchain-anthropic",
    "langchain-openai",
]


[tool.setuptools]
packages = ["src"]
